✨ feat: 렌더링 및 이벤트 핸들러 로직을 커스텀 훅으로 분리
- src/basic/hooks/useRendering.js 파일 생성
- performInitialRendering, setupEventHandlers 로직 분리 및 훅 구성
- main.basic.js에서 해당 훅을 import하여 main 함수 간결화
