+++ src/basic/hooks/useCart.js
@@ -1,11 +1,11 @@
 import { getCartItems } from '../components/CartDisplay.js';
 import { CartPriceUpdater } from '../services/CartPriceUpdater.js';
-import { 
-  getProductList, 
-  updateCartItems, 
-  updateCartItemCount, 
-  updateCartTotal, 
-  updateCartDiscount 
+import {
+  getProductList,
+  updateCartItems,
+  updateCartItemCount,
+  updateCartTotal,
+  updateCartDiscount,
 } from '../state/appState.js';
 
 /**
@@ -74,4 +74,4 @@ export function useCartPriceUpdate(cartDisp, discountCalculator, uiUpdater) {
   }
 
   return { doUpdatePricesInCart };
-} 
+}
